requirements.txt
contraption==0.0.8
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.15.4
pyparsing==2.3.1
python-dateutil==2.8.0
PyVISA==1.9.1
PyVISA-py==0.3.1
PyYAML==4.2b1
six==1.12.0
